chore/공통 컴포넌트를 atom 폴더에 넣기
