+++ src/components/atom/title/Title2.tsx
@@ -1,8 +1,8 @@
-import { TextProps } from '@components/title/Title1'
+import { TextProps } from '@components/atom/title/Title1'
 import { Text } from 'react-native'
 
-const Title3 = ({ text, ...props }: Readonly<TextProps>) => {
-  const size = 22
+const Title2 = ({ text, ...props }: Readonly<TextProps>) => {
+  const size = 26
 
   return (
     <Text
@@ -10,7 +10,7 @@ const Title3 = ({ text, ...props }: Readonly<TextProps>) => {
       className={props.className}
       style={[
         {
-          fontFamily: 'WantedSans-Medium',
+          fontFamily: 'WantedSans-SemiBold',
           fontSize: size,
           lineHeight: size * 1.5,
           letterSpacing: size * -0.025,
@@ -23,4 +23,4 @@ const Title3 = ({ text, ...props }: Readonly<TextProps>) => {
   )
 }
 
-export default Title3
+export default Title2
