+++ src/screens/Login/index.tsx
@@ -1,9 +1,9 @@
-import Body4 from '@components/body/Body4'
+import Body4 from '@components/atom/body/Body4'
 import { Image, Pressable, View } from 'react-native'
 import { SafeAreaView } from 'react-native-safe-area-context'
 import BackgroundImage from '../../../assets/images/login/background.svg'
 import MainPageBack from '@components/MainPageBack'
-import Body3 from '@components/body/Body3'
+import Body3 from '@components/atom/body/Body3'
 import KakaoIcon from '../../../assets/images/login/kakao.svg'
 import { NativeStackScreenProps } from '@react-navigation/native-stack'
 import { AuthStackParamList } from '@stackNav/Auth'
