+++ src/screens/NicknameWrite/index.tsx
@@ -4,9 +4,9 @@ import BackgroundImage from '../../../assets/images/login/background.svg'
 import MainPageBack from '@components/MainPageBack'
 import { NativeStackScreenProps } from '@react-navigation/native-stack'
 import { AuthStackParamList } from '@stackNav/Auth'
-import Title2 from '@components/title/Title2'
+import Title2 from '@components/atom/title/Title2'
 import { useState } from 'react'
-import Button from '@components/button/Button'
+import Button from '@components/atom/button/Button'
 import {
   launchImageLibraryAsync,
   MediaTypeOptions,
